requirements.txt
Django==1.8.2
argparse==1.2.1
django-model-utils==2.2
django-notifications-hq==0.7.0
jsonfield==1.0.3
six==1.9.0
wsgiref==0.1.2
